refactor: split firebase function call from firestore and spreadsheet
